+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Fixed
 - Show correct client in HostStatusPanel after full reload
+- Correctly refresh client configuration page after full reload
 
 ## [4.4.0.6] - 2024-09-05
 ### Fixed